fix(app-generic): circular dependency on subscriber pref use cases
