+++ libs/shared/src/entities/subscriber-preference/subscriber-preference.interface.ts
@@ -1,7 +1,29 @@
+import { ChannelTypeEnum, PreferenceOverrideSourceEnum } from '../../types';
+
 export interface IPreferenceChannels {
   email?: boolean;
   sms?: boolean;
   in_app?: boolean;
   chat?: boolean;
   push?: boolean;
+}
+
+export interface IPreferenceOverride {
+  channel: ChannelTypeEnum;
+  source: PreferenceOverrideSourceEnum;
+}
+
+export interface ISubscriberPreferenceResponse {
+  template: ITemplateConfiguration;
+  preference: {
+    enabled: boolean;
+    channels: IPreferenceChannels;
+    overrides: IPreferenceOverride[];
+  };
+}
+
+export interface ITemplateConfiguration {
+  _id: string;
+  name: string;
+  critical: boolean;
 }
