+++ packages/application-generic/src/usecases/get-subscriber-preference/get-subscriber-preference.usecase.ts
@@ -4,7 +4,11 @@ import {
   NotificationTemplateEntity,
   MemberRepository,
 } from '@novu/dal';
-import { ChannelTypeEnum, IPreferenceChannels } from '@novu/shared';
+import {
+  ChannelTypeEnum,
+  IPreferenceChannels,
+  ISubscriberPreferenceResponse,
+} from '@novu/shared';
 
 import { AnalyticsService } from '../../services';
 import { GetSubscriberPreferenceCommand } from './get-subscriber-preference.command';
@@ -60,24 +64,4 @@ export class GetSubscriberPreference {
       )
     );
   }
-}
-
-export interface ISubscriberPreferenceResponse {
-  template: ITemplateConfiguration;
-  preference: {
-    enabled: boolean;
-    channels: IPreferenceChannels;
-    overrides: IPreferenceOverride[];
-  };
-}
-
-export interface ITemplateConfiguration {
-  _id: string;
-  name: string;
-  critical: boolean;
-}
-
-export interface IPreferenceOverride {
-  channel: ChannelTypeEnum;
-  source: 'template' | 'subscriber';
 }
